Replace Markdown component with tsx
