--- src/components/Base/Markdown/index.ts
@@ -1,3 +1,0 @@
-import Markdown from './Markdown.vue'
-
-export default Markdown
